pyproject.toml
[tool.poetry]
name = "expense_api"
version = "0.1.0"
description = ""
authors = ["Isai Ramirez <isai.ramirez.stamaria15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
